SDK in target folder exists check, remove redundant path in zip_config, update the makefile

+++ go_script/src/mobile.ooyala.com/samples/ios/update_from_target_location/update_from_target_location.go
@@ -7,6 +7,7 @@ import "mobile.ooyala.com/common/util"
 import . "mobile.ooyala.com/common/path"
 import vc "mobile.ooyala.com/samples/config/vendor_config"
 import zc "mobile.ooyala.com/samples/config/zip_config"
+import "os"
 
 func run(fn func() error, l *gl.Logger) {
 	err := fn()
@@ -33,31 +34,32 @@ func main() {
 	zipConfig := zc.MakeiOSConfig(rootDir, l);
 
 	var configArgs args.Config
-	run(
-		func() error { 
-			var err error; 
-			configArgs, err = loadFlags(l); 
-			return err 
-		}, 
-	l)
+	run(func() error { var err error; configArgs, err = loadFlags(l); return err }, l)
 
 	sdkFolderPath := *configArgs.Path
 
-	removeOldOoyalaVendorFolders(config, l)
+	if sdkFolderPath == "" {
+		sdkFolderPath = "/Users/yigu/repos/ios-sdk/"
+	}
 
-	copyFromTargetFolders(config, zipConfig, sdkFolderPath, l)
+	// if the sdk from target location really exists
+	if exist, err := sdksExist(sdkFolderPath + zipConfig.CoreSDKFileNameStr, sdkFolderPath + zipConfig.FreewheelSDKFileNameStr, sdkFolderPath + zipConfig.IMASDKFileNameStr); exist == true && err == nil {
+		removeOldOoyalaVendorFolders(config, l)
 
-	unzipNewRCPackages(config, zipConfig, l)
+		copyFromTargetFolders(config, zipConfig, sdkFolderPath, l)
 
-	removeZipFiles(config, zipConfig, l)
+		unzipNewRCPackages(config, zipConfig, l)
 
-	//TEMPORARY: Remove all sample apps provided in the packages, until the sample apps are no longer included
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaCoreFolderPath, MakeFileName("SampleApps"))), l);
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaCoreFolderPath, MakeFileName("APIDocs"))), l);
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaFreewheelFolderPath, MakeFileName("FreewheelSampleApp"))), l);
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaFreewheelFolderPath, MakeFileName("APIDocs"))), l);
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaIMAFolderPath,MakeFileName("IMASampleApp"))), l);
-	util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaIMAFolderPath, MakeFileName("APIDocs"))), l);
+		removeZipFiles(config, zipConfig, l)
+
+		//TEMPORARY: Remove all sample apps provided in the packages, until the sample apps are no longer included
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaCoreFolderPath, MakeFileName("SampleApps"))), l);
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaCoreFolderPath, MakeFileName("APIDocs"))), l);
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaFreewheelFolderPath, MakeFileName("FreewheelSampleApp"))), l);
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaFreewheelFolderPath, MakeFileName("APIDocs"))), l);
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaIMAFolderPath,MakeFileName("IMASampleApp"))), l);
+		util.DeletePath(MakeFileAbs(Join(config.VendorOoyalaIMAFolderPath, MakeFileName("APIDocs"))), l);
+	}
 }
 
 func removeOldOoyalaVendorFolders(config vc.Config, l *log.Logger) {
@@ -70,17 +72,13 @@ func removeOldOoyalaVendorFolders(config vc.Config, l *log.Logger) {
 func copyFromTargetFolders(config vc.Config, zipConfig zc.Config, sdkFolderPath string, l *log.Logger) {
 	ol.ColorizedMethodPrintln(l)
 
-	if sdkFolderPath == "" {
-		sdkFolderPath = "/Users/yigu/repos/ios-sdk/"
-	}
-
-	err := util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.CoreSDKTargetPath + " " + zipConfig.CoreSDKFileNameStr}, l)
+	err := util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.CoreSDKFileNameStr + " " + zipConfig.CoreSDKFileNameStr}, l)
 	util.MaybeDie(err, l)
 
-	err = util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.FreewheelSDKTargetPath +  " " + zipConfig.FreewheelSDKFileNameStr}, l)
+	err = util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.FreewheelSDKFileNameStr +  " " + zipConfig.FreewheelSDKFileNameStr}, l)
 	util.MaybeDie(err, l)
 
-	err = util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.IMASDKTargetPath + " " + zipConfig.IMASDKFileNameStr}, l)
+	err = util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"cp " + sdkFolderPath + zipConfig.IMASDKFileNameStr + " " + zipConfig.IMASDKFileNameStr}, l)
 	util.MaybeDie(err, l)
 }
 
@@ -107,4 +105,19 @@ func removeZipFiles(config vc.Config, zipConfig zc.Config, l *log.Logger) {
 
 	err = util.RunBashCommandsInDir(config.VendorOoyalaRootFolderPath, []string{"rm " + zipConfig.IMASDKFileNameStr}, l)
 	util.MaybeDie(err, l)
+}
+
+func sdksExist(CoreSDKTargetPath string, FreewheelSDKTargetPath string, IMASDKTargetPath string) (bool, error) {
+    _, err := os.Stat(CoreSDKTargetPath)
+    if os.IsNotExist(err) { return false, nil }
+
+    _, err = os.Stat(FreewheelSDKTargetPath)
+    if os.IsNotExist(err) { return false, nil }
+
+    _, err = os.Stat(IMASDKTargetPath)
+    if os.IsNotExist(err) { return false, nil }
+
+    if err == nil { return true, nil }
+
+    return false, err
 }
